backend/requirements.txt
Django==5.2
django-ninja==1.4.2
gunicorn==23.0.0
whitenoise[brotli]==6.9.0
